requirements.txt
Pillow>=9.0.0
tk>=0.1.0
